Extract incremental sync logic into use case class
